move to ViewFileTreeStore.ts
